moved download logic to separate file
