+++ src/AppMain.js
@@ -3,34 +3,18 @@ import { Formik, Form } from 'formik';
 import JSZip from "jszip";
 import { Divider, Button } from '@material-ui/core';
 import { MetaFields } from './FieldComponents/FieldGroups';
-
-//todo replace with webextensions
-function download(filename, base64) {
-  var element = document.createElement('a');
-  element.setAttribute('href', base64);
-  element.setAttribute('download', filename);
-
-  element.style.display = 'none';
-  document.body.appendChild(element);
-
-  element.click();
-
-  document.body.removeChild(element);
-}
+import download from './download';
 
 const handleSubmit = (values, { setSubmitting }) => {
-  var zip = new JSZip();
   const pack = values.pack_mcmeta;
   setTimeout(() => {
-    zip.file('pack.mcmeta', JSON.stringify({ pack }, null, 2));
-    zip.folder('data');
-    zip.generateAsync({type:"base64"}).then(function (base64) {
-      download(values.pack_name, base64);
+    generateZip(pack).then(function (base64) {
+      onSuccess(values.pack_name, base64);
       setSubmitting(false);
-  }, function (err) {
-      alert(err);
-  });
-  }, 400);
+    }, function (err) {
+      onError(err);
+    });
+  })
 };
 const AppMain = () =>{
     return (
@@ -50,9 +34,9 @@ const AppMain = () =>{
          render={({ isSubmitting }) => (
           <Form className="form-container">
             <MetaFields />
-            <Divider variant="middle"></Divider>
+            <Divider component="div" variant="middle"></Divider>
             <Button variant="outlined" download="foo-bar.zip" type="submit" disabled={isSubmitting}>
-              Submit
+              Download DataPack
             </Button>
           </Form>
         )}>
@@ -61,3 +45,18 @@ const AppMain = () =>{
 }
     
 export default AppMain;
+
+const onError = (err) => {
+  alert(err);
+}
+
+const onSuccess = (pack_name, base64) => {
+  download(pack_name, base64);
+}
+function generateZip(pack) {
+  var zip = new JSZip();
+  zip.file('pack.mcmeta', JSON.stringify({ pack }, null, 2));
+  zip.folder('data');
+  return zip.generateAsync({ type: "base64" });
+}
+
